updated CMAKE_CXX_COMPILER_ID check for GNU instead of GCC
this fixed a build issue we experienced, caused by filesystem check not including -std=c++17 on amazon linux 2 when other dependencies are included along with Drogon in the same project.

GCC is not listed as one of the possible values in CMake documentation. GNU is listed as one of the possible values and it is what is set in our build environment.